@@ -6,41 +6,29 @@ using CommandLine;
 
 namespace CleanEmptyFolders
 {
-	class MainClass
+	static class MainClass
 	{
-		private static List<string> emptyDirs = new List<string>();
+		private static List<string> emptyDirs = [];
 
 		public static int Main(string[] args)
 		{
 			Console.WriteLine("Clean Empty folders");
 
 			var result = Parser.Default.ParseArguments<Options>(args);
-			var exitCode = result.MapResult(options =>
-											DoIt(options), errors =>
-											{
-
-												Console.WriteLine(errors);
-												return 1;
-											});
+			var exitCode = result.MapResult(DoIt, errors => 
+													{
+														Console.WriteLine(errors);
+														return 1;
+													});
 			Console.WriteLine("Clean Empty folders finished with exit code " + exitCode);
 			return exitCode;
 		}
 
 		static int DoIt(Options options)
 		{
-			if (options.Verbose) Console.WriteLine("Starting folder: {0}", options.StartingFolder);
 
-			if (!Directory.Exists(options.StartingFolder))
-			{
-				Console.WriteLine("Folder '{0}' does not exist.", options.StartingFolder);
-				return 1;
-			}
-			else
-			{
-				Console.WriteLine("Starting in '{0}'.", options.StartingFolder);
-			}
 
-			emptyDirs = new List<string>();
+			emptyDirs = [];
 			ProcessDir(options.StartingFolder);
 
 			foreach (var item in emptyDirs)
@@ -71,18 +59,14 @@ namespace CleanEmptyFolders
 
 			var files = Directory.GetFiles(dir);
 
-			if (dirs.Length == 0 && (FilesAreUnimportant(files)))
+			if (dirs.Length == 0 && FilesAreUnimportant(files))
 				emptyDirs.Add(dir);
 
 		}
 
-		private static bool FilesAreUnimportant(string[] files)
-		{
-			return !files.Any() || files.Any() && files.All(FileIsUnimportant);
+        private static bool FilesAreUnimportant(string[] files) => files.Length == 0 || files.Length != 0 && files.All(FileIsUnimportant);
 
-		}
-
-		static bool FileIsUnimportant(string item)
+        static bool FileIsUnimportant(string item)
 		{
 			if (item.EndsWith(".AAE", StringComparison.InvariantCultureIgnoreCase))
 				return true;
